+++ packages/scenes/src/components/VizPanel/VizPanel.tsx
@@ -324,6 +324,17 @@ export class VizPanel<TOptions = {}, TFieldConfig extends {} = {}> extends Scene
       }),
     };
 
+    if (this._dataWithFieldConfig.annotations) {
+      this._dataWithFieldConfig.annotations = applyFieldOverrides({
+        data: this._dataWithFieldConfig.annotations,
+        fieldConfig: this.state.fieldConfig,
+        fieldConfigRegistry,
+        replaceVariables: this.interpolate,
+        theme: config.theme2,
+        timeZone: rawData.request?.timezone,
+      });
+    }
+
     if (!pluginDataSupport.alertStates) {
       this._dataWithFieldConfig.alertState = undefined;
     }