+++ packages/scenes/src/variables/VariableDependencyConfig.test.ts
@@ -1,8 +1,13 @@
+import { VariableRefresh } from '@grafana/schema';
 import { SceneObjectBase } from '../core/SceneObjectBase';
 import { SceneObjectState } from '../core/types';
 
 import { VariableDependencyConfig } from './VariableDependencyConfig';
 import { ConstantVariable } from './variants/ConstantVariable';
+import { TestVariable } from './variants/TestVariable';
+import { TestObjectWithVariableDependency, TestScene } from './TestScene';
+import { SceneVariableSet } from './sets/SceneVariableSet';
+import { SceneTimeRange } from '../core/SceneTimeRange';
 
 interface TestState extends SceneObjectState {
   query: string;
@@ -24,7 +29,7 @@ class TestObj extends SceneObjectBase<TestState> {
   }
 }
 
-describe('VariableDependencySet', () => {
+describe('VariableDependencyConfig', () => {
   it('Should be able to extract dependencies from all state', () => {
     const sceneObj = new TestObj();
     const deps = new VariableDependencyConfig(sceneObj, {});
@@ -77,10 +82,70 @@ describe('VariableDependencySet', () => {
     const fn = jest.fn();
     const deps = new VariableDependencyConfig(sceneObj, { onReferencedVariableValueChanged: fn });
 
-    deps.variableUpdatesCompleted(new Set([new ConstantVariable({ name: 'not-dep', value: '1' })]));
+    deps.variableUpdateCompleted(new ConstantVariable({ name: 'not-dep', value: '1' }), true);
     expect(fn.mock.calls.length).toBe(0);
 
-    deps.variableUpdatesCompleted(new Set([new ConstantVariable({ name: 'queryVarA', value: '1' })]));
+    deps.variableUpdateCompleted(new ConstantVariable({ name: 'queryVarA', value: '1' }), true);
     expect(fn.mock.calls.length).toBe(1);
+  });
+
+  describe('Should remember when an object is waiting for variables', () => {
+    it('Should notify as soon as next variable completes', async () => {
+      const A = new TestVariable({
+        name: 'A',
+        query: 'A.*',
+        value: '',
+        text: '',
+        options: [],
+        refresh: VariableRefresh.onTimeRangeChanged,
+      });
+      const B = new TestVariable({ name: 'B', query: '$A', value: '', text: '', options: [] });
+      const C = new TestVariable({ name: 'C', query: '$B', value: '', text: '', options: [] });
+
+      const nestedObj = new TestObjectWithVariableDependency({ title: '$C' });
+      const set = new SceneVariableSet({ variables: [A, B, C] });
+      const timeRange = new SceneTimeRange();
+      const scene = new TestScene({
+        $variables: set,
+        $timeRange: timeRange,
+        nested: nestedObj,
+      });
+
+      scene.activate();
+      nestedObj.activate();
+
+      nestedObj.doSomethingThatRequiresVariables();
+
+      // Verify testObj has not done anything yet (still waiting for variables)
+      expect(nestedObj.state.didSomethingCount).toBe(0);
+
+      A.signalUpdateCompleted();
+      B.signalUpdateCompleted();
+      C.signalUpdateCompleted();
+
+      // Now it can
+      expect(nestedObj.state.didSomethingCount).toBe(1);
+
+      // Do something while no variables are loading
+      nestedObj.doSomethingThatRequiresVariables();
+      expect(nestedObj.state.didSomethingCount).toBe(2);
+
+      // change time range to trigger A loading
+      timeRange.onRefresh();
+      expect(A.state.loading).toBe(true);
+
+      // Now do something and it should wait
+      nestedObj.doSomethingThatRequiresVariables();
+      expect(nestedObj.state.didSomethingCount).toBe(2);
+
+      // B completes
+      A.signalUpdateCompleted();
+
+      // No change in value so B should not be loading
+      expect(B.state.loading).toBe(false);
+
+      // No need to wait now as no dependency loading
+      expect(nestedObj.state.didSomethingCount).toBe(3);
+    });
   });
 });