+++ packages/scenes/src/variables/adhoc/AdHocFiltersSet.tsx
@@ -11,6 +11,8 @@ import { ControlsLabel } from '../../utils/ControlsLabel';
 import { AdHocFilterRenderer } from './AdHocFilterRenderer';
 import { AdHocFilterBuilder } from './AdHocFilterBuilder';
 import { css } from '@emotion/css';
+import { sceneGraph } from '../../core/sceneGraph';
+import { DataQueryExtended, SceneQueryRunner } from '../../querying/SceneQueryRunner';
 
 export interface AdHocFilterSetState extends SceneObjectState {
   /** Defaults to Filters */
@@ -143,7 +145,8 @@ export class AdHocFilterSet extends SceneObjectBase<AdHocFilterSetState> {
     }
 
     const otherFilters = this.state.filters.filter((f) => f.key !== currentKey).concat(this.state.baseFilters!);
-    let keys = await ds.getTagKeys({ filters: otherFilters });
+    const queries = this._getSceneQueries();
+    let keys = await ds.getTagKeys({ filters: otherFilters, queries });
 
     if (override) {
       keys = keys.concat(override.values);
@@ -175,7 +178,9 @@ export class AdHocFilterSet extends SceneObjectBase<AdHocFilterSetState> {
 
     // Filter out the current filter key from the list of all filters
     const otherFilters = this.state.filters.filter((f) => f.key !== filter.key).concat(this.state.baseFilters!);
-    let values = await ds.getTagValues({ key: filter.key, filters: otherFilters });
+
+    const timeRange = sceneGraph.getTimeRange(this).state.value;
+    let values = await ds.getTagValues({ key: filter.key, filters: otherFilters, timeRange });
 
     if (override) {
       values = values.concat(override.values);
@@ -193,6 +198,29 @@ export class AdHocFilterSet extends SceneObjectBase<AdHocFilterSetState> {
       label: value,
       value,
     }));
+  }
+
+  /**
+   * Get all queries in the scene that have the same datasource as this AdHocFilterSet
+   */
+  private _getSceneQueries(): DataQueryExtended[] {
+    const runners = sceneGraph.findAllObjects(
+      this.getRoot(),
+      (o) => o instanceof SceneQueryRunner
+    ) as SceneQueryRunner[];
+
+    const applicableRunners = runners.filter((r) => r.state.datasource?.uid === this.state.datasource?.uid);
+
+    if (applicableRunners.length === 0) {
+      return [];
+    }
+
+    const result: DataQueryExtended[] = [];
+    applicableRunners.forEach((r) => {
+      result.push(...r.state.queries);
+    });
+
+    return result;
   }
 }